# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/v1/reservations": {
      "post": {
        "operationId": "ReservationsController_create_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReservationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Reservation created successfully"
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Customer or room not found"
          },
          "409": {
            "description": "Room not available for selected dates"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new reservation",
        "tags": [
          "Reservations"
        ]
      },
      "get": {
        "operationId": "ReservationsController_findAll_v1",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Items per page",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 10,
              "type": "number"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "Sort field",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "description": "Sort order",
            "schema": {
              "default": "desc",
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Filter by status",
            "schema": {
              "type": "string",
              "enum": [
                "PENDING",
                "CONFIRMED",
                "CHECKED_IN",
                "CHECKED_OUT",
                "CANCELLED",
                "NO_SHOW"
              ]
            }
          },
          {
            "name": "customerId",
            "required": false,
            "in": "query",
            "description": "Filter by customer ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "roomId",
            "required": false,
            "in": "query",
            "description": "Filter by room ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "checkInFrom",
            "required": false,
            "in": "query",
            "description": "Filter by check-in date from",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "checkInTo",
            "required": false,
            "in": "query",
            "description": "Filter by check-in date to",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "checkOutFrom",
            "required": false,
            "in": "query",
            "description": "Filter by check-out date from",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "checkOutTo",
            "required": false,
            "in": "query",
            "description": "Filter by check-out date to",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of reservations"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all reservations with pagination and filters",
        "tags": [
          "Reservations"
        ]
      }
    },
    "/api/v1/v1/reservations/today/arrivals": {
      "get": {
        "operationId": "ReservationsController_getTodayArrivals_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of today arrivals"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get today arrivals",
        "tags": [
          "Reservations"
        ]
      }
    },
    "/api/v1/v1/reservations/today/departures": {
      "get": {
        "operationId": "ReservationsController_getTodayDepartures_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of today departures"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get today departures",
        "tags": [
          "Reservations"
        ]
      }
    },
    "/api/v1/v1/reservations/code/{code}": {
      "get": {
        "operationId": "ReservationsController_findByCode_v1",
        "parameters": [
          {
            "name": "code",
            "required": true,
            "in": "path",
            "description": "Reservation code",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Reservation details"
          },
          "404": {
            "description": "Reservation not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get reservation by code",
        "tags": [
          "Reservations"
        ]
      }
    },
    "/api/v1/v1/reservations/{id}": {
      "get": {
        "operationId": "ReservationsController_findOne_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Reservation UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Reservation details"
          },
          "404": {
            "description": "Reservation not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get reservation by ID",
        "tags": [
          "Reservations"
        ]
      },
      "patch": {
        "operationId": "ReservationsController_update_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Reservation UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateReservationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Reservation updated successfully"
          },
          "400": {
            "description": "Cannot update reservation"
          },
          "404": {
            "description": "Reservation not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update reservation",
        "tags": [
          "Reservations"
        ]
      },
      "delete": {
        "operationId": "ReservationsController_remove_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Reservation UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Reservation deleted successfully"
          },
          "404": {
            "description": "Reservation not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete reservation (soft delete)",
        "tags": [
          "Reservations"
        ]
      }
    },
    "/api/v1/v1/reservations/{id}/confirm": {
      "post": {
        "operationId": "ReservationsController_confirm_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Reservation UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Reservation confirmed"
          },
          "400": {
            "description": "Cannot confirm reservation"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Confirm reservation",
        "tags": [
          "Reservations"
        ]
      }
    },
    "/api/v1/v1/reservations/{id}/check-in": {
      "post": {
        "operationId": "ReservationsController_checkIn_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Reservation UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckInDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Check-in completed"
          },
          "400": {
            "description": "Cannot check in"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Check-in guest",
        "tags": [
          "Reservations"
        ]
      }
    },
    "/api/v1/v1/reservations/{id}/check-out": {
      "post": {
        "operationId": "ReservationsController_checkOut_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Reservation UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckOutDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Check-out completed"
          },
          "400": {
            "description": "Cannot check out"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Check-out guest",
        "tags": [
          "Reservations"
        ]
      }
    },
    "/api/v1/v1/reservations/{id}/cancel": {
      "post": {
        "operationId": "ReservationsController_cancel_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Reservation UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Reservation cancelled"
          },
          "400": {
            "description": "Cannot cancel reservation"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Cancel reservation",
        "tags": [
          "Reservations"
        ]
      }
    },
    "/api/v1/v1/reservations/{id}/no-show": {
      "post": {
        "operationId": "ReservationsController_markNoShow_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Reservation UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Reservation marked as no-show"
          },
          "400": {
            "description": "Cannot mark as no-show"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Mark reservation as no-show",
        "tags": [
          "Reservations"
        ]
      }
    },
...
    "/api/v1/v1/services/reservation/{id}": {
      "delete": {
        "operationId": "ServicesController_removeFromReservation_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Reservation service UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Service removed from reservation"
          },
          "404": {
            "description": "Reservation service not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Remove service from reservation",
        "tags": [
          "Additional Services"
        ]
      }
    },
    "/api/v1/v1/services/reservation/{reservationId}": {
      "get": {
        "operationId": "ServicesController_getReservationServices_v1",
        "parameters": [
          {
            "name": "reservationId",
            "required": true,
            "in": "path",
            "description": "Reservation UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Reservation services"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get services for a reservation",
        "tags": [
          "Additional Services"
        ]
      }
    },
...
  },
  "info": {
    "title": "Hotel Management API",
    "description": "API for hotel management system",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      }
    },
    "schemas": {
...
      "CreateReservationDto": {
        "type": "object",
        "properties": {
          "customerId": {
            "type": "string",
            "description": "Customer ID"
          },
          "roomId": {
            "type": "string",
            "description": "Room ID"
          },
          "checkInDate": {
            "type": "string",
            "example": "2024-12-01",
            "description": "Check-in date"
          },
          "checkOutDate": {
            "type": "string",
            "example": "2024-12-05",
            "description": "Check-out date"
          },
          "adults": {
            "type": "number",
            "example": 2,
            "default": 1,
            "description": "Number of adults"
          },
          "children": {
            "type": "number",
            "example": 0,
            "default": 0,
            "description": "Number of children"
          },
          "notes": {
            "type": "string",
            "description": "Notes"
          }
        },
        "required": [
          "customerId",
          "roomId",
          "checkInDate",
          "checkOutDate"
        ]
      },
      "UpdateReservationDto": {
        "type": "object",
        "properties": {
          "checkInDate": {
            "type": "string",
            "example": "2024-12-01",
            "description": "Check-in date"
          },
          "checkOutDate": {
            "type": "string",
            "example": "2024-12-05",
            "description": "Check-out date"
          },
          "adults": {
            "type": "number",
            "example": 2,
            "default": 1,
            "description": "Number of adults"
          },
          "children": {
            "type": "number",
            "example": 0,
            "default": 0,
            "description": "Number of children"
          },
          "notes": {
            "type": "string",
            "description": "Notes"
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "CONFIRMED",
              "CHECKED_IN",
              "CHECKED_OUT",
              "CANCELLED",
              "NO_SHOW"
            ],
            "description": "Reservation status"
          }
        }
      },
      "CheckInDto": {
        "type": "object",
        "properties": {
          "notes": {
            "type": "string",
            "description": "Notes for check-in"
          }
        }
      },
      "CheckOutDto": {
        "type": "object",
        "properties": {
          "notes": {
            "type": "string",
            "description": "Notes for check-out"
          }
        }
      },
...
    }
  }
}